Markers component split into Markers and NewMarkers, All components briefly cleaned up to get rid of some warnings
