+++ src/Components/Filters.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import {
   LayersControl,
   FeatureGroup,
