+++ src/Components/Maps.jsx
@@ -1,8 +1,6 @@
-import React from "react";
-import {
-  TileLayer,
-  LayersControl,
-} from "react-leaflet";
+import { TileLayer, LayersControl } from "react-leaflet";
+
+
 //Each section is a different map type
 export default function MapTiles() {
   return (
