+++ src/Components/Markers.jsx
@@ -1,65 +1,20 @@
-import React, { useState } from "react";
 import useAxios from "axios-hooks";
-import { useMapEvents, Marker, Popup } from "react-leaflet";
-import MarkerForm from "./MarkerForm";
+import { Marker, Popup } from "react-leaflet";
 import PopupDisplay from "./PopupDisplay";
-import Modal from "@material-ui/core/Modal";
-import { makeStyles } from "@material-ui/core/styles";
-import Fade from "@material-ui/core/Fade";
 
-function rand() {
-  return Math.round(Math.random() * 20) - 10;
-}
 
-function getModalStyle() {
-  const top = 50 + rand();
-  const left = 50 + rand();
-
-  return {
-    top: `${top}%`,
-    left: `${left}%`,
-    transform: `translate(-${top}%, -${left}%)`,
-  };
-}
-
-const useStyles = makeStyles((theme) => ({
-  modal: {
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  paper: {
-    backgroundColor: theme.palette.background.paper,
-    border: "2px solid #000",
-    boxShadow: theme.shadows[5],
-    padding: theme.spacing(2, 4, 3),
-  },
-}));
-
+//Recieves Marker data from database
 export default function Markers() {
-  const [markers, setMarkers] = useState([]);
-
-  const classes = useStyles();
-  const [modalStyle] = React.useState(getModalStyle);
-  const [modal, setModal] = useState(false);
-  const map = useMapEvents({
-    dblclick(e) {
-      const newMarker = e.latlng;
-      setModal(true);
-      setMarkers([...markers, newMarker]);
-    },
-  });
-
+  
   const [{ data, loading, error }] = useAxios(
     "https://angler-reg-api.herokuapp.com/pins"
   );
   if (loading) return <p>Loading...</p>;
   if (error) return <p>Error!</p>;
   const returnData = data.pins;
-  console.log("ReturnData",returnData)
+
   return (
     <>
-      {/* needs a unique id */}
       {returnData.map((marker) => (
         <Marker key={marker.id}position={[marker.location.x, marker.location.y]}>
           <Popup>
@@ -72,17 +27,6 @@ export default function Markers() {
               rating={marker.rating}
             />
           </Popup>
-          <Modal
-            open={modal}
-            className={classes.modal}
-            onClose={() => setModal(false)}
-          >
-            <Fade in={modal}>
-              <div className={classes.paper}>
-                <MarkerForm marker={marker} onClose={() => setModal(false)} />
-              </div>
-            </Fade>
-          </Modal>
         </Marker>
       ))}
     </>
